Unify install bookinfo and install/upgrade smcp test
